Update schema to use decimals and adjust models and controllers accordingly

+++ db/migrate/20190518162204_create_transactions.rb
@@ -3,7 +3,7 @@ class CreateTransactions < ActiveRecord::Migration[5.2]
     create_table :transactions do |t|
       t.string :counterparty
       t.datetime :date
-      t.integer :amount
+      t.decimal :amount
 
       t.belongs_to :account, index: true, foreign_key: true
