+++ db/migrate/20190601022656_add_accounttotal_to_accounts.rb
@@ -1,5 +1,5 @@
 class AddAccounttotalToAccounts < ActiveRecord::Migration[5.2]
   def change
-    add_column :accounts, :account_total, :integer, default: 0
+    add_column :accounts, :account_total, :decimal, default: 0
   end
 end
